+++ REvoDesign/common/MultiMutantDesigner.py
@@ -14,15 +14,14 @@ from REvoDesign.tools.pymol_utils import (
     is_distal_residue_pair,
 )
 
-from REvoDesign.tools.mutant_tools import (
-    extract_mutants_from_mutant_id,
-)
+from REvoDesign.tools.mutant_tools import extract_mutant_from_pymol_object
 
 
 class MultiMutantDesigner:
-    def __init__(self, molecule, chain_id):
+    def __init__(self, molecule, chain_id, sequence):
         self.molecule = molecule
         self.chain_id = chain_id
+        self.sequence = sequence
         self.cmap = 'bwr_r'
         self.total_design_cases = 20
 
@@ -34,7 +33,9 @@ class MultiMutantDesigner:
         # Initialize mutant tree for design
         _mutant_tree = {
             group_id: {
-                mutant_id: extract_mutants_from_mutant_id(mutant_string=mutant_id)[1]
+                mutant_id: extract_mutant_from_pymol_object(
+                    pymol_object=mutant_id, sequence=self.sequence
+                )
                 for mutant_id in cmd.get_object_list(f'({group_id})')
             }
             for group_id in cmd.get_names(type='group_objects', enabled_only=1)